Use int64_t for version values in internal shape-inference plumbing (#8149)
### Description

`ModelProto::ir_version()` and `OperatorSetIdProto::version()` are
`int64_t`, but internal plumbing carried these version values as `int`.
This widens the **non-exported** carriers to `int64_t` so they match the
proto source:

- the `ir_version` parameter/field threaded through `graphProtoToGraph`
/ `convertAttribute(s)` (`ir_pb_converter.cc`), `InferShapesImpl`, and
`GraphInferenceContext`;
- the local `domain_to_opset_version` map in
`OpSchema::GetNodeDeterminism`.

**Public API/ABI is intentionally preserved.** `OperatorSetVersion`,
`OpSchema::SinceVersion`, and the pure-virtual
`ISchemaRegistry::GetSchema(int)` are left as `int`. The one narrowing
that is genuinely forced by the public `GetSchema(int)` boundary is now
an explicit `static_cast<int>` there, instead of an implicit truncation
at the map assignment.

### Scope / safety

- Change is contained to `ir_pb_converter.cc`, `implementation.{cc,h}`,
and `schema.cc`; no other TU constructs `GraphInferenceContext` or calls
the (static) `graphProtoToGraph`.
- `ir_version` is only ever compared (`>= 4`), never fed into an `int`
sink after widening.
- Verified: the edited TUs plus every TU referencing these symbols
compile clean under the full `clang-tidy` config.

### Motivation and Context

Type-consistency cleanup: internal version values now match the
`int64_t` proto fields they come from. Opset/IR versions are small, so
this is not a bug fix — it removes silent `int64_t`→`int` truncations in
internal plumbing without touching the public schema API.

+++ onnx/common/ir_pb_converter.cc
@@ -17,7 +17,7 @@
 namespace ONNX_NAMESPACE {
 
 // Part 1: convert ONNX Protobuf to IR
-static std::unique_ptr<Graph> graphProtoToGraph(const GraphProto& gp, bool nested, int ir_version = IR_VERSION);
+static std::unique_ptr<Graph> graphProtoToGraph(const GraphProto& gp, bool nested, int64_t ir_version = IR_VERSION);
 
 static Tensor tensorProtoToTensor(const ONNX_NAMESPACE::TensorProto& tp) {
   Tensor ret;
@@ -119,7 +119,7 @@ static Tensor tensorProtoToTensor(const ONNX_NAMESPACE::TensorProto& tp) {
   return ret;
 }
 
-static void convertAttribute(const ONNX_NAMESPACE::AttributeProto& ap, Node& n, const int ir_version = IR_VERSION) {
+static void convertAttribute(const ONNX_NAMESPACE::AttributeProto& ap, Node& n, const int64_t ir_version = IR_VERSION) {
   Symbol sym = Symbol(ap.name());
   switch (ap.type()) {
     case ONNX_NAMESPACE::AttributeProto_AttributeType_FLOAT:
@@ -204,7 +204,7 @@ static void convertAttribute(const ONNX_NAMESPACE::AttributeProto& ap, Node& n, 
   }
 }
 
-static void convertAttributes(const ONNX_NAMESPACE::NodeProto& np, Node& n, const int ir_version = IR_VERSION) {
+static void convertAttributes(const ONNX_NAMESPACE::NodeProto& np, Node& n, const int64_t ir_version = IR_VERSION) {
   for (int i = 0; i < np.attribute_size(); i++) {
     convertAttribute(np.attribute(i), n, ir_version);
   }
@@ -239,7 +239,7 @@ static Value* createDummyValue(
   return v;
 }
 
-std::unique_ptr<Graph> graphProtoToGraph(const ONNX_NAMESPACE::GraphProto& gp, bool nested, const int ir_version) {
+std::unique_ptr<Graph> graphProtoToGraph(const ONNX_NAMESPACE::GraphProto& gp, bool nested, const int64_t ir_version) {
   auto g = std::make_unique<Graph>();
 
   if (gp.has_name()) {
